requirements/container_app_requirements.txt
psycopg2-binary>=2.9.10
transformers>=4.52.4
Pillow>=11.2.1
torch>=2.7.1
detoxify>=0.5.2
protobuf>=6.31.0
accelerate>=1.8.0
azure-functions>=1.23.0
torchvision>=0.22.1
sentencepiece>=0.2.0,<1.0.0
azure-identity>=1.23,<2.0.0
azure-keyvault-secrets>=4.10.0,<5.0.0
openai>=1.90.0,<2.0.0
fastapi>=0.116.0,<1.0.0
requests
gunicorn
uvicorn
python-multipart